+++ line-config-ui/package.js
@@ -1,12 +1,12 @@
 Package.describe({
   name: 'storyteller:line-config-ui',
   summary: 'Blaze configuration templates for LINE OAuth.',
-  version: '1.0.1',
+  version: '1.1.0',
   git: 'https://github.com/StorytellerCZ/meteor-accounts-line'
 });
 
 Package.onUse(api => {
-  api.versionsFrom('1.8.3');
+  api.versionsFrom('1.9');
   api.use('ecmascript', 'client');
   api.use('templating@1.3.2', 'client');