+++ accounts-line/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.2 - 2021-02-14
+* Fix link to package in notice
+* Bump minimum Meteor version to 1.9
+
 ## 1.0.1 - 2020-01-01
 * Fix imports
 * Update dependencies