+++ line-oauth/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.0
+* Bump minimum Meteor version to 1.9
+* Replace HTTP package with Fetch package
+
 ## 1.0.1 - 2020-01-01
 * Update dependencies
 * Update minimum Meteor version to 1.8.3